Refactor iterators to add a way to get size of data
Previously our iter.Iter structs only referred to iterators returned by
the go-tsz packages. The problem we're trying to solve is that from the
go-tsz iterators there's no way to get the length of the underlying
data, but we need to know that to limit the size of the chunk cache.
This patch makes structs that implement the mdata.Store interface return
lists of iterator generators. These can be cached and reused, which is
what we need for the chunk cache.

The commit also includes some changes done by gofmt.