Introduces a second chunk format to include length
By default all new chunks will be written in the new format. This means
the second byte of the binary data is a uint8 that specifies how many
10min intervals this chunk is long.

We only specify with a precision of 10min to require less space to
define the length.

If a configured chunk span (or aggmetric chunk span) is not dividable
by 10min we error at startup.